# api/docs/swagger.yaml
basePath: /api/v1/
definitions:
...
  domain.StudentInfoShort:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  domain.Survey:
    properties:
      questions:
        items:
          $ref: '#/definitions/domain.SurveyQuestion'
        type: array
      required:
        type: boolean
      title:
        type: string
    type: object
  domain.SurveyAnswer:
    properties:
      answer:
        type: string
      questionId:
        type: string
    type: object
  domain.SurveyQuestion:
    properties:
      answerOptions:
        items:
          type: string
        type: array
      answerType:
        type: string
      id:
        type: string
      question:
        type: string
    type: object
  domain.SurveyResult:
    properties:
      answers:
        items:
          $ref: '#/definitions/domain.SurveyAnswer'
        type: array
      id:
        type: string
      moduleId:
        type: string
      student:
        $ref: '#/definitions/domain.StudentInfoShort'
      submittedAt:
        type: string
    type: object
...
  v1.createSurveyInput:
    properties:
      questions:
        items:
          $ref: '#/definitions/v1.question'
        type: array
      required:
        type: boolean
      title:
        type: string
    required:
    - title
    type: object
  v1.dataResponse:
    properties:
      count:
        type: integer
      data:
        type: object
    type: object
...
  v1.question:
    properties:
      answerOptions:
        items:
          type: string
        type: array
      answerType:
        type: string
      question:
        type: string
    required:
    - answerType
    - question
    type: object
...
  v1.response:
    properties:
      message:
        type: string
    type: object
...
  v1.uploadResponse:
    properties:
      url:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: REST API for UltraThreads App
  title: UltraThreads API
  version: "1.0"
paths:
...
  /admins/modules/{id}/survey:
    delete:
      consumes:
      - application/json
      description: admin delete survey
      parameters:
      - description: module id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Survey'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: Admin Delete Survey
      tags:
      - admins-surveys
    get:
      consumes:
      - application/json
      description: admin get survey
      parameters:
      - description: module id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Survey'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: Admin Get Survey
      tags:
      - admins-surveys
    post:
      consumes:
      - application/json
      description: admin create/update survey
      parameters:
      - description: module id
        in: path
        name: id
        required: true
        type: string
      - description: survey info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.createSurveyInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: Admin Create/Update Survey
      tags:
      - admins-surveys
  /admins/modules/{id}/survey/results:
    get:
      consumes:
      - application/json
      description: admin get all survey results
      parameters:
      - description: skip
        in: query
        name: skip
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: module id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.dataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: Admin Get Survey Results
      tags:
      - admins-surveys
  /admins/modules/{id}/survey/results/{studentId}:
    get:
      consumes:
      - application/json
      description: admin get all survey results
      parameters:
      - description: module id
        in: path
        name: id
        required: true
        type: string
      - description: student id
        in: path
        name: studentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SurveyResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: Admin Get Survey Results
      tags:
      - admins-surveys
...
  /admins/upload/file:
    post:
      consumes:
      - multipart/form-data
      description: admin upload file
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.uploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: Admin upload File
      tags:
      - admins-upload
  /admins/upload/image:
    post:
      consumes:
      - multipart/form-data
      description: admin upload image
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.uploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: Admin upload Image
      tags:
      - admins-upload
  /admins/upload/video:
    post:
      consumes:
      - multipart/form-data
      description: admin upload video
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.uploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: Admin upload Video
      tags:
      - admins-upload
...
securityDefinitions:
  AdminAuth:
    in: header
    name: Authorization
    type: apiKey
  StudentsAuth:
    in: header
    name: Authorization
    type: apiKey
  UsersAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"